requirements.txt
discord.py
xdice
python-dateutil
